Remove SuiteRun
